+++ utils.h
@@ -3,12 +3,12 @@
 #include <cstdlib>
 
 //inline
-double sigmoid(double input);
+float sigmoid(float input);
 
 //returns a random float between zero and 1
 //inline
-double randFloat(); //	   {return (rand())/(RAND_MAX+1.0);}
+float randFloat(); //	   {return (rand())/(RAND_MAX+1.0);}
 
 //returns a random float in the range -1 < n < 1
 //inline
-double randomClamped(); //   {return randFloat() - randFloat();}
+float randomClamped(); //   {return randFloat() - randFloat();}
